Refactor booking module: Remove search-related methods and prepare for search module
