+++ backend/src/config/swagger.config.ts
@@ -6,7 +6,7 @@ export const swaggerConfig = new DocumentBuilder()
     .setDescription('VinShuttle service API documentation')
     .setVersion('1.0')
     .addTag('auth', 'Authentication')
-    .addTag('booking', 'booking management')
+    .addTag('search', 'search vehicle for each service')
     .addTag('users', 'User management')
     .addTag('vehicles', 'Vehicle management')
     .addTag('vehicle-categories', 'Vehicle category management')
