+++ backend/src/modules/search/search.di-token.ts
@@ -1,0 +1,1 @@
+export const SEARCH_SERVICE = Symbol('SearchService');
