+++ backend/src/modules/search/search.module.ts
@@ -1,13 +1,11 @@
 import { Module } from "@nestjs/common";
-import { MongooseModule } from "@nestjs/mongoose";
-import { BookingController } from "src/modules/booking/booking.controller";
-import { BOOKING_SERVICE } from "src/modules/booking/booking.di-token";
-import { Booking, BookingSchema } from "src/modules/booking/booking.schema";
-import { BookingService } from "src/modules/booking/booking.service";
 import { DriverScheduleModule } from "src/modules/driver-schedule/driver-schedule.module";
 import { KeytokenModule } from "src/modules/keytoken/keytoken.module";
 import { PricingModule } from "src/modules/pricing/pricing.module";
 import { ScenicRouteModule } from "src/modules/scenic-route/scenic-route.module";
+import { SearchController } from "src/modules/search/search.controller";
+import { SEARCH_SERVICE } from "src/modules/search/search.di-token";
+import { SearchService } from "src/modules/search/search.service";
 import { TripModule } from "src/modules/trip/trip.module";
 import { VehicleCategoryModule } from "src/modules/vehicle-categories/vehicle-category.module";
 import { VehiclesModule } from "src/modules/vehicles/vehicles.module";
@@ -15,8 +13,8 @@ import { ShareModule } from "src/share/share.module";
 
 const dependencies = [
     {
-        provide: BOOKING_SERVICE,
-        useClass: BookingService
+        provide: SEARCH_SERVICE,
+        useClass: SearchService
     },
     // {
     //     provide: BOOKING_REPOSITORY
@@ -26,12 +24,6 @@ const dependencies = [
 
 @Module({
     imports: [
-        MongooseModule.forFeature([
-            {
-                name: Booking.name,
-                schema: BookingSchema
-            }
-        ]),
         VehiclesModule,
         VehicleCategoryModule,
         DriverScheduleModule,
@@ -41,9 +33,9 @@ const dependencies = [
         ShareModule,
         KeytokenModule
     ],
-    controllers: [BookingController],
+    controllers: [SearchController],
     providers: [...dependencies],
     exports: [...dependencies]
 }
 )
-export class BookingModule { }
+export class SearchModule { }
